Rename files from cosign to sigstore
... and minimal other changes to keep tests passing.

Note that test fixtures are usually not renamed. That's primarily
because some of them contain cosign repository names, and besides
many were actually cosign-created anyway.
